requirements.txt
async-timeout==4.0.2
attrs==24.2.0
backoff==2.1.2
blinker==1.5
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
docopt==0.6.2
ecs-logging==2.0.0
elastic-apm==6.12.0
elastic-opentelemetry==0.8.1
exceptiongroup==1.2.2
Flask==2.2.2
googleapis-common-protos==1.69.2
grpcio==1.68.1
h11==0.14.0
idna==3.4
importlib_metadata==8.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-system-metrics
opentelemetry-instrumentation-wsgi
opentelemetry-proto==1.31.1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-resourcedetector-gcp==1.8.0a0
opentelemetry-sdk==1.31.1
opentelemetry-sdk-extension-aws==2.1.0
opentelemetry-semantic-conventions
opentelemetry-util-http
outcome==1.3.0.post0
packaging==21.3
pipreqs==0.4.11
protobuf==5.29.4
psutil==5.9.2
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
structlog==22.1.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.21.0